# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.AdminUserDTO:
    properties:
      authType:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      hierarchyLevel:
        type: string
      id:
        type: string
      reportsTo:
        type: string
      teamIds:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.CreateUserRequest:
    properties:
      authType:
        description: '"local" (default) or "sso"'
        enum:
        - local
        - sso
        type: string
      email:
        type: string
      fullName:
        type: string
      hierarchyLevel:
        type: string
      id:
        description: Optional - will be auto-generated from username if not provided
        type: string
      password:
        description: Required for local users, omit for SSO users
        type: string
      reportsTo:
        type: string
      username:
        description: Required - used to generate ID if not provided
        type: string
    required:
    - email
    - fullName
    - hierarchyLevel
    - username
    type: object
...
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.PaginationDTO:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.UpdateUserRequest:
    properties:
      authType:
        enum:
        - local
        - sso
        type: string
      email:
        type: string
      fullName:
        type: string
      hierarchyLevel:
        type: string
      password:
        type: string
      reportsTo:
        type: string
      username:
        type: string
    type: object
...
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.UserLiteDTO:
    properties:
      fullName:
        type: string
      hierarchyLevel:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.UsersLiteResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.UserLiteDTO'
        type: array
    type: object
  github_com_agopalakrishnan_teams360_backend_interfaces_dto.UsersResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.PaginationDTO'
      users:
        items:
          $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.AdminUserDTO'
        type: array
    type: object
...
info:
  contact: {}
  description: Team360 Health Check Application REST API. Assesses team health across
    11 dimensions with hierarchical org support.
  title: Team360 API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      description: Returns a paginated, filterable list of users. Requires admin (level-1)
        role.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 25, max 100)
        in: query
        name: pageSize
        type: integer
      - description: Search by username/name/email
        in: query
        name: search
        type: string
      - description: Filter by hierarchy level ID
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.UsersResponse'
        "500":
          description: Failed to query users
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users (paginated)
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      description: Creates a new user (local or SSO auth type). Local users require
        a password. Requires admin (level-1) role.
      parameters:
      - description: User to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.AdminUserDTO'
        "400":
          description: Invalid request body or missing/short password for local user
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse'
        "500":
          description: Failed to hash password or create user
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - admin-users
  /admin/users/{id}:
    delete:
      description: Deletes a user. Requires admin (level-1) role.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.MessageResponse'
        "500":
          description: Failed to delete user
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - admin-users
    put:
      consumes:
      - application/json
      description: Updates a user's profile, role, auth type, and/or password. Re-derives
        supervisor chains if reportsTo or hierarchy level changes. Requires admin
        (level-1) role.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.AdminUserDTO'
        "400":
          description: Invalid request body, invalid authType, cannot set password
            for SSO user, or password required when switching to local auth
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse'
        "500":
          description: Failed to hash password, update user, or update password
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - admin-users
  /admin/users/lite:
    get:
      description: Returns minimal user data (id, username, full name, hierarchy level)
        for the full user set, for dropdowns/pickers. Requires admin (level-1) role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.UsersLiteResponse'
        "500":
          description: Failed to query users
          schema:
            $ref: '#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users (lite)
      tags:
      - admin-users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"